Check to see if there are any imported torrents before removing them

+++ scripts/clear_imported_torrents.sh
@@ -1,5 +1,13 @@
 #!/usr/bin/env sh
 
 cd /Users/lopopolo/Downloads
-rm *.torrent.imported
+
+shopt -s nullglob
+found=0
+for i in *.torrent.imported; do
+  found=1
+done
+shopt -u nullglob
+
+[ $found -eq 1 ] && rm *.torrent.imported
